LM reconstruction is working and is able to calculate the sensitivity image and I am working to be able to apply the normalisation factors.
I had problems with the RelatedBins class, therefore I created several
fucntion which take as input a vector<Bin> and a DataSymmetriesFromBin,
which essentially is the same thing ( at least for this application).

+++ src/include/stir/recon_buildblock/PoissonLogLikelihoodWithLinearModelForMeanAndListModeData.h
@@ -122,7 +122,7 @@ protected:
   //! \author Nikos Efthimiou
   //! \details This is part of some functionality I transfer from lm_to_projdata.
   //! The total number of events to be *STORED* not *PROCESSED*.
-  int num_events_to_store;
+  unsigned long int num_events_to_store;
 
   //!
    //! \brief do_time_frame
